kadai complete

+++ app/Http/Controllers/tasksController.php
@@ -6,6 +6,8 @@ use Illuminate\Http\Request;
 
 use App\Task;
 
+use App\Http\Controllers\Controller;
+
 class TasksController extends Controller
 {
     /**
@@ -14,26 +16,42 @@ class TasksController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function index()
-    {
-        $tasks = Task::all();
+    {   
+        
+        if (\Auth::check())
+        {
+        $user = \Auth::user();
+        $tasks = Task::all() ->where("user_id", "=", $user ->id);
+        
         return view('tasks.index', [
         'tasks' => $tasks,
                 ]);
-    }
-
+        }
+        else
+        {
+            return view ('welcome');
+        }
+}
     /**
      * Show the form for creating a new resource.
      *
      * @return \Illuminate\Http\Response
      */
     public function create()
-    {
+    {   
+         if (\Auth::check())
+        {
         $task = new Task;
         return view('tasks.create', [
         'task' => $task,
     ]);
         
-    }
+        }
+       else
+        {
+            return view ('welcome');
+        }
+}
 
     /**
      * Store a newly created resource in storage.
@@ -51,6 +69,7 @@ class TasksController extends Controller
         $task = new Task;
         $task->status = $request->status;
         $task->content = $request->content;
+        $task->user_id = \Auth::user()->id;
         $task->save();
         return redirect('/');
     }
@@ -63,12 +82,28 @@ class TasksController extends Controller
      */
     public function show($id)
     {   
+        if (\Auth::check())
+        {
         $task = Task::find($id);
+        
+        if (\Auth::user()->id === $task->user_id)
+        {
         return view('tasks.show', [
         'task' => $task,
             ]);
-    }
+          }   
+          
+          else
+          {
+              return redirect('/');
+          }
+        }
 
+        else
+        {
+            return view('welcome');
+        }
+    }
     /**
      * Show the form for editing the specified resource.
      *
@@ -76,11 +111,27 @@ class TasksController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function edit($id)
-    {
+    {   
+        if(\Auth::check())
+        {
         $task = Task::find($id);
+        
+        if (\Auth::user()->id === $task->user_id)
+        {
         return view('tasks.edit', [
         'task' => $task,
                 ]);
+        }
+        
+        else
+        {
+            return redirect('/');
+        }
+       }  
+        else
+        {
+            return view('welcome');
+        }
     }
 
     /**
@@ -92,6 +143,7 @@ class TasksController extends Controller
      */
     public function update(Request $request, $id)
     {   
+       
         $this->validate($request, [
         'status' => 'required|max:10',
         'content' => 'required|max:191',
@@ -102,9 +154,10 @@ class TasksController extends Controller
         $task->status = $request->status;
         $task->content = $request->content;
         $task->save();
-        
+        if (\Auth::user()->id === $task->user_id){
             return redirect('/');
     }
+}
 
     /**
      * Remove the specified resource from storage.
@@ -113,9 +166,13 @@ class TasksController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function destroy($id)
-    {
+    {   
         $task = Task::find($id);
+        
+        if (\Auth::user()->id === $task->user_id)
+        {
         $task->delete();
         return redirect('/');
+        }
     }
 }
